requirements.txt
# Core Dependencies
numpy==1.24.3
pandas==2.0.3
pymongo==4.5.0
kafka-python==2.0.2
redis==5.0.0
pyspark==3.5.0

# Data Generation
Faker==19.6.2
scipy==1.11.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Optional (for local development)
jupyter==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
